checker done

+++ libft/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: msodor <msodor@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/12 09:04:12 by msodor            #+#    #+#             */
-/*   Updated: 2022/10/12 09:51:09 by msodor           ###   ########.fr       */
+/*   Updated: 2023/02/12 08:02:49 by msodor           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_strlen(const char *str)
+int	ft_strlen(const char *str)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (str[i])
